# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgroLink API
  version: 1.0.0
  description: |
    AgroLink connects farmers, buyers, and delivery agents through a unified platform.
    This specification documents the public REST endpoints available from the Flask backend.
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
  - url: https://agrolink-backend-nezp.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
    AuthToken:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [farmer, buyer, delivery_agent, superadmin]
        phone:
          type: string
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum: [farmer, buyer, delivery_agent]
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/login:
    post:
      summary: Obtain a JWT access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AuthToken'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/profile:
    get:
      summary: Retrieve the profile of the authenticated user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
security:
  - bearerAuth: []